+++ retesteth/EthChecks.cpp
@@ -4,6 +4,7 @@
 #include <retesteth/TestOutputHelper.h>
 #include <iostream>
 #include <csignal>
+#include <thread>
 
 namespace test {
 
@@ -18,7 +19,7 @@ void eth_require(bool _flag)
     if (!_flag)
     {
         std::raise(SIGABRT);
-        TestOutputHelper::get().markError();
+        TestOutputHelper::get().markError("Flag error");
     }
 }
 
@@ -27,25 +28,29 @@ void eth_check_message(bool _flag, std::string const& _message)
     if (!_flag)
     {
         std::cerr << _message << std::endl;
-        TestOutputHelper::get().markError();
+        TestOutputHelper::get().markError(_message);
     }
 }
 
 void eth_require_message(bool _flag, std::string const& _message)
 {
     if (!_flag)
-    {
-        std::cerr << _message << std::endl;
-        std::raise(SIGABRT);
-        TestOutputHelper::get().markError();
-    }
+        eth_fail(_message);
 }
 
 void eth_fail(std::string const& _message)
 {
     std::cerr << _message << std::endl;
+    std::this_thread::sleep_for(std::chrono::seconds(1));
     std::raise(SIGABRT);
-    TestOutputHelper::get().markError();
+    TestOutputHelper::get().markError(_message);
+    throw std::exception();
+}
+
+void eth_error(std::string const& _message)
+{
+    std::cerr << _message << std::endl;
+    TestOutputHelper::get().markError(_message);
 }
 
 }
